refactor remaining supporting functions
